@@ -1,4 +1,4 @@
-// swift-tools-version: 5.9
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -20,7 +20,7 @@ let package = Package(
         .target(
             name: "ZTronVideoPlayer",
             swiftSettings: [
-                .enableExperimentalFeature("StrictConcurrency=minimal")
+                .enableExperimentalFeature("StrictConcurrency=complete")
             ]
         ),
         
